fix: no notice on config errors
Catching the exception to decide creation method also caught config
errors.

For example:
```
- !LinearController
  foo: 0.9
- !CpuPool
```

The daemon for this configuration ran but did not create the `CpuPool`
and displayed no error message for the invalid argument `foo` for
`LinearController`. After this patch running the config example above
throws an exception which greatly helps to detect typos.